fix: 修复backoff策略构建时multiplier类型不匹配问题
将make_backoff_policy中的multiplier参数转换为f64类型以匹配接口要求
更新版本号至0.1.9

+++ src/downloader.rs
@@ -46,7 +46,7 @@ pub async fn make_backoff_policy(max_timeout_secs: u64) -> ExponentialBackoff {
     let mut building = ExponentialBackoffBuilder::default();
     let policy = building
         .with_initial_interval(Duration::from_secs(initial_interval))
-        .with_multiplier(multiplier)
+        .with_multiplier(multiplier as f64)
         .with_randomization_factor(0.2)
         .with_max_elapsed_time(Some(Duration::from_secs(max_elapsed_time)));
     policy.build()
